Changed to double

+++ C/src/CHeterodyning.c
@@ -1,9 +1,9 @@
 #include "CHeterodyning.h"
 
-extern float data [SAMPLE_COUNT];
-extern float carrier[SAMPLE_COUNT];
+extern double data [SAMPLE_COUNT];
+extern double carrier[SAMPLE_COUNT];
 
-float result [SAMPLE_COUNT];
+double result [SAMPLE_COUNT];
 
 int main(int argc, char**argv){
     printf("Running Unthreaded Test\n");
